refactor: 레이어간 의존성 방향 규칙을 지키기 위해 store context를 shared로 내림
