+++ src/features/playlist/edit/ui/form.component.tsx
@@ -1,9 +1,9 @@
 import { useCallback } from 'react';
 import { SubmitHandler } from 'react-hook-form';
-import { useStores } from '@/app/_providers/stores.context';
 import { PlaylistForm, PlaylistFormProps, PlaylistFormValues } from '@/entities/playlist';
 import { Playlist } from '@/shared/api/http/types/playlists';
 import { useI18n } from '@/shared/lib/localization/i18n.context';
+import { useStores } from '@/shared/lib/store/stores.context';
 import { useDialog } from '@/shared/ui/components/dialog';
 import { useUpdatePlaylist } from '../api/use-update-playlist.mutation';
