+++ src/shared/lib/store/stores.context.tsx
@@ -1,14 +1,23 @@
 'use client';
 
 import { createContext, useContext } from 'react';
-import { StoreApi, UseBoundStore } from 'zustand';
-import { UIState } from '@/entities/ui-state';
-import { CurrentPartyroom } from 'entities/current-partyroom';
 
-export type Stores = {
-  useUIState: UseBoundStore<StoreApi<UIState.Model>>;
-  useCurrentPartyroom: UseBoundStore<StoreApi<CurrentPartyroom.Model>>;
-};
+/**
+ * context 구현부에서 declaration-merging을 통해 인터페이스를 확장하여 사용해주세요.
+ *
+ * @see https://www.typescriptlang.org/docs/handbook/declaration-merging.html
+ *
+ * @example
+ * ```ts
+ * declare module '@/shared/lib/store/stores.context' {
+ *   interface Stores {
+ *     useStoreA: ...;
+ *     useStoreB: ...;
+ *   }
+ * }
+ * ```
+ */
+export interface Stores {}
 
 export const StoresContext = createContext<Stores | null>(null);
