+++ src/widgets/my-playlist/ui/my-playlist.component.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useState } from 'react';
-import { useStores } from '@/app/_providers/stores.context';
 import { AddPlaylistButton } from '@/features/playlist/add';
 import { AddMusicsToPlaylistButton } from '@/features/playlist/add-musics';
 import { Playlists, EditablePlaylists, PlaylistListItem } from '@/features/playlist/list';
@@ -8,6 +7,7 @@ import { RemovePlaylistButton } from '@/features/playlist/remove';
 import { Playlist } from '@/shared/api/http/types/playlists';
 import { useI18n } from '@/shared/lib/localization/i18n.context';
 import { replaceVar } from '@/shared/lib/localization/split-render';
+import { useStores } from '@/shared/lib/store/stores.context';
 import { Drawer, DrawerProps } from '@/shared/ui/components/drawer';
 import { TextButton } from '@/shared/ui/components/text-button';
 import { PFArrowLeft } from '@/shared/ui/icons';
